Rename remap apply module
