+++ uxarray/remap/__init__.py
@@ -1,6 +1,6 @@
+from .apply_weights import _apply_weights
 from .inverse_distance_weighted import _inverse_distance_weighted_remap
 from .nearest_neighbor import _nearest_neighbor_remap
-from .precomputed import _apply_weights
 from .weights import RemapWeights, load_remap_weights
 
 __all__ = (
